requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
strawberry-graphql==0.230.3